miniblog/requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
dominate==2.9.1
Faker==24.3.0
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.2
